refactor(macos-diag): extract ProfileDrilldown for reuse
Carve the per-profile expanded body (metadata grid + payloads list with
parsed settings table) out of MacosDiagProfilesTab into a standalone
ProfileDrilldown component so the upcoming JAMF Profiles tab can render
identical drilldown UI without duplicating ~140 lines of JSX and ~20
makeStyles entries.

The drilldown reads logListFontSize directly from useUiStore so callers
just pass `profile`. Re-exported from the workspace index for cross-
workspace consumption.

Zero behavior change in the macos-diag tab.
